DB_SIZE_CHECKIN/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
mysql-connector-python>=8.3.0
